requirements.txt
gitdb==4.0.10
GitPython==3.1.31
mdurl==0.1.2
pandas==1.5.3
Pillow==9.4.0
requests==2.28.1
rich==13.3.4
streamlit==1.21.0
